crear_solicitudes_proyecto/urls.py
"""crear_solicitudes_proyecto URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/2.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from django.shortcuts import redirect
from rest_framework_jwt.views import obtain_jwt_token, refresh_jwt_token

from solicitudes.views import current_user

urlpatterns = [
    path('', lambda request: redirect('api/v1/solicitudes/', permanent=False)),
    path('admin/', admin.site.urls),
    path('api/v1/', include('solicitudes.urls')),
    path('api-auth/', include('rest_framework.urls')),
    path(r'api-token-auth/', obtain_jwt_token),
    path(r'api-token-refresh/', refresh_jwt_token),
    path('api/v1/rest-auth/', include('rest_auth.urls')),
    path('api/v1/rest-auth/registration/', include('rest_auth.registration.urls')),
    path('current_user/', current_user),
]
