requirements.txt
djangorestframework-jwt
Django==2.2.5
djangorestframework==3.10.3
django-rest-auth==0.9.5
django-allauth==0.40.0
django-cors-headers==3.1.0
boto3==1.9.243
awscli==1.16.253
